+++ CHANGELOG.md
@@ -1,4 +1,12 @@
-# Version 0.4.0 (current)
+# Version 0.4.1 (current)
+###07.08.2019
+
+[+] added stop mutex for Sniffer, DispatcherBase, WrappedListener, GPIOSPIPortBase and Connection
+[+] added fileSize func for IO
+[*] fixed bug with pcap_next for Sniffer
+[*] fixed bug in DispatcherBase for unhandled events
+
+# Version 0.4.0
 [*] bugfixes
 [*] optimization